use long tensor instead of byte

+++ torch_datasets/samplers/balanced_batch_sampler.py
@@ -59,7 +59,7 @@ class BalancedBatchSampler(torch.utils.data.sampler.BatchSampler):
 
         # Randomly shuffle batch ids
         batch_ids = random.sample(batch_ids, self.batch_size)
-        batch_ids = torch.ByteTensor(batch_ids)
+        batch_ids = torch.LongTensor(batch_ids)
 
         return batch_ids
